api/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[argon2]
argon2-cffi
httpx
python-multipart
email-validator
icalendar>=5.0.11